wip: move plugin manager code from settings to pluginuimanager module
